plexus_aws_infrastructure/plexus_aws_infrastructure_stack.py
from aws_cdk import (
    # Duration,
    Stack,
    # aws_sqs as sqs,
)
from constructs import Construct

class PlexusAwsInfrastructureStack(Stack):

    def __init__(self, scope: Construct, construct_id: str, **kwargs) -> None:
        super().__init__(scope, construct_id, **kwargs)

        # The code that defines your stack goes here

        # example resource
        # queue = sqs.Queue(
        #     self, "PlexusAwsInfrastructureQueue",
        #     visibility_timeout=Duration.seconds(300),
        # )
